Proposed fix for ring.GetReplicationSetForOperation().
Problem: GetReplicationSetForOperation returns wrong
ReplicationSet.MaxUnavailableZones when there are fewer total zones in
the ring than the replication factor.

This is a proposed fix and probaby needs additional work.

In the changes in the tests, the modified RF=3 ones show the cases where
the previous iplementation was not working as expected.